+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 3.9.11
+- Fixed subscription logic
+- Added getAllResults method
+- Updated subscribe method
+
 # 3.9.10
 -   Moved TrackData to constant.dart to ensure it is publicly accessible and can be imported correctly without relying on internal paths.
 - - Updated project functionality to align with version 0.7.1 change log: